neovim -> nvf migration
Mostly done. Stuff is not working entirely as before
when it comes to autocompletion, but there's a singular
file instead of a thousand, and even though I have no
idea how blink-cmp works (not that I had any idea about
how the cmp I was using before worked), I'm possitive
that I will eventually get the hang of it.

Oh, also moved from fastfetch to microfetch, because
why not.